food-service/requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary
requests==2.31.0
python-decouple==3.8
django_filter
drf_spectacular